+++ SpatialGDK/Source/SpatialGDK/Private/EngineClasses/SpatialNetDriver.cpp
@@ -1647,12 +1647,17 @@ void USpatialNetDriver::PollPendingLoads()
 
 void USpatialNetDriver::TickFlush(float DeltaTime)
 {
-	PollPendingLoads();
-
-	// Super::TickFlush() will not call ReplicateActors() because Spatial connections have InternalAck set to true.
-	// In our case, our Spatial actor interop is triggered through ReplicateActors() so we want to call it regardless.
-
 	const USpatialGDKSettings* SpatialGDKSettings = GetDefault<USpatialGDKSettings>();
+
+	if (SpatialGDKSettings->bRunSpatialWorkerConnectionOnGameThread)
+	{
+		if (Connection != nullptr && Connection->IsConnected())
+		{
+			Connection->QueueLatestOpList();
+		}
+	}
+
+	PollPendingLoads();
 
 	if (IsServer() && GetSpatialOSNetConnection() != nullptr && PackageMap->IsEntityPoolReady() && bIsReadyToStart)
 	{
@@ -1696,6 +1701,16 @@ void USpatialNetDriver::TickFlush(float DeltaTime)
 
 	TimerManager.Tick(DeltaTime);
 
+	if (SpatialGDKSettings->bRunSpatialWorkerConnectionOnGameThread)
+	{
+		if (Connection != nullptr && Connection->IsConnected())
+		{
+			Connection->ProcessOutgoingMessages();
+		}
+	}
+
+	// Super::TickFlush() will not call ReplicateActors() because Spatial connections have InternalAck set to true.
+	// In our case, our Spatial actor interop is triggered through ReplicateActors() so we want to call it regardless.
 	Super::TickFlush(DeltaTime);
 }